pyproject.toml
[project]
name = "fastapi-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
